+++ CHANGELOG.md
@@ -4,8 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog][1], and this project adheres to [Semantic Versioning][2].
 
 ## [Unreleased]
+### Added
+ - `Instant` archetype supporting basic instant/amount arithmetics.
+
 ### Fixed
- - Make newtypes implement Sync even if markers do not implement it.
+ - Newtypes implement Sync even if the markers does not implement it.
 
 ## [0.1.0] 2019-11-06
 ### Added